+++ example/app/.devconfig.ts
@@ -1,4 +1,4 @@
-import type { ServiceConfig } from '@tnotifier/devon';
+import type { ServiceConfig } from '@evilkiwi/devon';
 
 export const config: ServiceConfig = {
     // Put your env, docker-compose etc. config here!