+++ packages/devon/src/helpers/compose/index.ts
@@ -2,7 +2,6 @@ import { dump } from 'js-yaml';
 import type { ServiceConfig } from '../../types';
 import { getServices } from '../services';
 
-// TODO: Typing docker-compose format?
 export const compileCompose = async (configs: Record<string, ServiceConfig>) => {
     const { definition } = await getServices();
     const version = definition.composeVersion ?? '3.9';