+++ packages/devon/src/templates/services/redis.ts
@@ -1,4 +1,4 @@
-export const redis = `import type { ServiceConfig } from '@tnotifier/devon';
+export const redis = `import type { ServiceConfig } from '@evilkiwi/devon';
 
 export const config: ServiceConfig = {
     compose: {